Export component and functions in different files
